third_party/blink/web_tests/accessibility/aria-none-role.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body id="body">

<h3 role="none">
<a href="#">Link</a> and text
</h3>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the aria 'none' role works by successfully ignoring the element in the AX tree.");

    if (window.accessibilityController) {

          var body = document.getElementById("body");
          body.focus();

          var firstChild = accessibilityController.focusedElement.childAtIndex(0);
          shouldBeTrue("firstChild.isIgnored");
    }

</script>

</body>
</html>
