third_party/blink/web_tests/accessibility/aria-none-role-expected.txt

Link and text

This tests that the aria 'none' role works by successfully ignoring the element in the AX tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS firstChild.isIgnored is true
PASS successfullyParsed is true

TEST COMPLETE
